inline training function, and other review responses
