+++ external/fv3fit/fv3fit/keras/__init__.py
@@ -1,4 +1,3 @@
 from ._models.models import History
 from ._models.recurrent import PureKerasModel, StepwiseModel
-from ._training import get_model
 from . import _models
